Change some fields from int to bigint/long

+++ tskswp-rest-api/src/main/kotlin/com/tasksweeper/entities/AccountEntity.kt
@@ -7,7 +7,7 @@ object Account : Table("tskswp.account") {
     val username = varchar("username", 256)
     val email = varchar("email", 256)
     val password = varchar("password", 256)
-    val level = integer("level")
+    val level = long("level")
     override val primaryKey = PrimaryKey(username, name = "user_pkey")
 }
 
@@ -15,5 +15,5 @@ data class AccountDTO(
     val username: String,
     val email: String,
     @JsonIgnore val password: String,
-    val level: Int
+    val level: Long
 )
